+++ ignite/config/plugins/config.go
@@ -15,8 +15,9 @@ import (
 type Config struct {
 	path string
 
-	// Plugins holds the list of installed plugins.
-	Plugins []Plugin `yaml:"plugins"`
+	// Apps holds the list of installed Ignite Apps.
+	// Ignite Apps are implemented as plugins.
+	Apps []Plugin `yaml:"apps"`
 }
 
 // Plugin keeps plugin name and location.
@@ -43,7 +44,7 @@ type Plugin struct {
 	With map[string]string `yaml:"with,omitempty"`
 
 	// Global holds whether the plugin is installed globally
-	// (default: $HOME/.ignite/plugins/plugins.yml) or locally for a chain.
+	// (default: $HOME/.ignite/apps/igniteapps.yml) or locally for a chain.
 	Global bool `yaml:"-"`
 }
 
@@ -135,7 +136,7 @@ func (c *Config) Save() error {
 // Returns also true if there's a local plugin with the module name equal to
 // that path.
 func (c Config) HasPlugin(path string) bool {
-	return slices.ContainsFunc(c.Plugins, func(cp Plugin) bool {
+	return slices.ContainsFunc(c.Apps, func(cp Plugin) bool {
 		if cp.HasPath(path) {
 			return true
 		}